+++ CHANGELOG.md
@@ -15,6 +15,7 @@ Release TBD
 - Use normal names for default campaigns and categories.
 - Let Campaigns control whether ads are shown to owners, admins, etc.
 - Refactor url validations, now done through separate call to admin index.php.
+- Implement GD-based image resizing to support transparent PNG.
 
 ## Version 0.3.0
 Release 2017-10-06